requirements.txt
customtkinter
tknav
mysql-connector-python
uuid
bcrypt